Move Godot kotlin specific files out of jni/
