+++ src/lifecycle/class_loader.cpp
