+++ src/lifecycle/platforms/jvm_loader_ios.cpp
